requirements.txt
ffmpeg-python
numpy
mutagen
pywhispercpp